refactor(help): renaming and getting rid of utils
