+++ src/app/components/components/help/help.component.html
@@ -41,23 +41,23 @@
         <td-highlight lang="typescript">
             <![CDATA[
           import {
-            IHelpMenuDataItem,
-            DraggableHelpWindowDialogService
+            IHelpItem,
+            HelpWindowService
           } from '@covalent/help';
 
-          oneItem: IHelpMenuDataItem[] = [
+          oneItem: IHelpItem[] = [
             {
               url: 'https://github.com/Teradata/covalent/blob/develop/README.md',
             },
           ];
 
-          itemWithRawMarkdown: IHelpMenuDataItem[] = [
+          itemWithRawMarkdown: IHelpItem[] = [
             {
               markdownString: '# Heading',
             },
           ];
 
-          multipleItems: IHelpMenuDataItem[] = [
+          multipleItems: IHelpItem[] = [
             {
               url: 'https://raw.githubusercontent.com/Teradata/covalent-code-editor/master/docs/API.md',
               title: 'Code Editor API',
@@ -68,14 +68,14 @@
             },
           ];
 
-          oneItemWithAnchor: IHelpMenuDataItem[] = [
+          oneItemWithAnchor: IHelpItem[] = [
             {
               url: 'https://raw.githubusercontent.com/Teradata/covalent/develop/docs/DEVELOPER_GUIDE.md',
               anchor: 'Adding a new documentation component',
             },
           ];
 
-          nestedItems: IHelpMenuDataItem[] = [
+          nestedItems: IHelpItem[] = [
             {
               title: 'Covalent Components',
               children: [
@@ -92,7 +92,7 @@
             },
           ];
 
-          options: { name: string; value: IHelpMenuDataItem[] }[] = [
+          options: { name: string; value: IHelpItem[] }[] = [
             {
               name: 'One item',
               value: this.oneItem,
@@ -106,15 +106,15 @@
             { name: 'Nested items', value: this.nestedItems },
           ];
 
-          selected: { name: string; value: IHelpMenuDataItem[] } = this.options[0];
-          currentItems: IHelpMenuDataItem[] = this.selected.value;
+          selected: { name: string; value: IHelpItem[] } = this.options[0];
+          currentItems: IHelpItem[] = this.selected.value;
           userInput: string = this.prettyJson(this.currentItems);
 
           select(): void {
             this.use(this.selected.value);
           }
 
-          use(items: IHelpMenuDataItem[]): void {
+          use(items: IHelpItem[]): void {
             this.currentItems = items;
             this.userInput = this.prettyJson(this.currentItems);
           }
@@ -123,7 +123,7 @@
             this.use(JSON.parse(this.userInput));
           }
 
-          prettyJson(items: IHelpMenuDataItem[]): string {
+          prettyJson(items: IHelpItem[]): string {
             return JSON.stringify(items, undefined, 4);
           }
             ]]>
@@ -135,7 +135,7 @@
 </mat-card>
 
 <mat-card>
-  <mat-card-title>Draggable Help Window Dialog Service</mat-card-title>
+  <mat-card-title>Help Window Service</mat-card-title>
   <mat-card-subtitle>
     A service that opens a HelpWindowComponent inside a draggable dialog. Uses the openDraggable method of the TdDialogService.
   </mat-card-subtitle>
@@ -162,26 +162,26 @@
         <td-highlight lang="typescript">
           <![CDATA[
           import {
-            IHelpMenuDataItem,
-            DraggableHelpWindowDialogService,
+            IHelpItem,
+            HelpWindowService,
             HelpWindowComponent,
           } from '@covalent/help';
 
           windowOpen: boolean = false;
           ref: MatDialogRef<HelpWindowComponent>;
-          currentItems: IHelpMenuDataItem[] = [
+          currentItems: IHelpItem[] = [
           {
             url: 'https://github.com/Teradata/covalent/blob/develop/README.md',
           },
           ];
 
-          constructor(private draggableHelpWindowDialogService: DraggableHelpWindowDialogService) {}
+          constructor(private _helpWindowService: HelpWindowService) {}
 
           openDialog(): void {
             if (this.windowOpen) {
               this.closeDialog();
             }
-            this.ref = this.draggableHelpWindowDialogService.open({items: this.currentItems});
+            this.ref = this._helpWindowService.open({items: this.currentItems});
             this.ref.afterOpened().subscribe(() => {
               this.windowOpen = true;
             });
