+++ src/platform/help/public-api.ts
@@ -1,5 +1,4 @@
 export * from './help.module';
 export * from './help.component';
 export * from './help-window/help-window.component';
-export * from './draggable-help-window-dialog/draggable-help-window-dialog.service';
-export * from './help.utils';
+export * from './help-window-service/help-window.service';
